+++ src/databar/__init__.py
@@ -99,7 +99,7 @@ from .models import (
     Folder,
 )
 
-__version__ = "2.2.0"
+__version__ = "2.3.0"
 __all__ = [
     "DatabarClient",
     # exceptions